+++ src/main/java/com/gromed/demo/repository/PresentationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PresentationRepository extends JpaRepository<Presentation, Integer> {
+public interface PresentationRepository extends JpaRepository<Presentation, Long> {
 }
